+++ migrations/postgres/20251112180821_extend_dataset_statistics_with_object_links.sql
@@ -1,0 +1,7 @@
+/* ------------------------------ */
+
+ALTER TABLE dataset_statistics
+ADD COLUMN num_object_links BIGINT NOT NULL DEFAULT 0,
+ADD COLUMN object_links_size BIGINT NOT NULL DEFAULT 0;
+
+/* ------------------------------ */